+++ test/clj_kondo/condition_always_true_test.clj
@@ -18,6 +18,10 @@
                (defn bar [x] (if x 2 inc))"
               '{:linters {:condition-always-true {:level :warning}}})))
   (assert-submaps2
-   []
+   [{:file "<stdin>",
+     :row 1,
+     :col 7,
+     :level :warning,
+     :message "Condition always true"}]
    (lint! "(when #'inc 2)"
           '{:linters {:condition-always-true {:level :warning}}})))